Added keyboard support!
ChromaMidiMPad and ChromaMidiKeyb are separate programs, one for a chroma keyboard and one for a chroma mousepad. I figured separating them this way would be easier and could allow uses to have their mousepads and keyboards on different midi channels/ports. The mousepad version is unchanged from last commit, but the keyboard version is still somewhat experimental.
Razer simplifies all of their keyboards down into a 5x21 matrix of LED's, so some lights will be repeated.